Add support for Jupyter notebook input file
Turns out it was actually simple to read in notebooks, convert them to
Markdown and then carry on with normal execution. The tricky part was
getting the output images, which have to be saved separately. For now,
the only way to set the HTML title for notebooks is through a JSON data
file.