Add workflow to run tests on all OS & Python (#7)
By tests I really mean build the docs at this point. Tests will come eventually. 
Since tests were failing for Python < 3.9 and Windows, this also fixes those issues.
For Python <3.9 the `|` operator for dicts is not available. 
So use a dictionary unpacking solution instead. 
For Windows, it was a matter of explicitly setting the encoding to utf-8 when 
reading and writing text files.