Use pypa build to create the distributions
Add the required configuration to use the `build` tool instead of
calling setup.py, which is now not really needed (but will be kept for
compatibility). Update the CI configuration to capture the dependencies
form requirements files so that we can build and test on different
versions of the packages.