Add list of siblings for each page
The `page["siblings"]` list contains all pages (the full dict, not just
the ID) that share the same folder with it excluding itself. This is
useful for creating index pages (like a list of blog posts). Had to
split the final rendering from the MD->HTML conversion so that templates
can access the sibling "body" elements (e.g., for summaries).