Use markdown-it-py to render Markdown content
Replace the MyST parser, which actually uses markdown-it-py. MyST is
meant to be used with sphinx and didn't really offer that much in the
way of customizing the Markdown rendering process. With markdown-it-py
we can use some of their plugins and even write some of our own to
support things like syntax highlighting for code blocks (with language
specification, like in GitHub flavored markdown). For now, just replace
the rendering and add a few plugins (typographic substitutions,
footnotes, tables, and anchors for headings.