Mechanism for adding data to a single page
Data in yml or json files was being propagated to every page in a
folder. This is useful but, particularly for notebooks, we may want to
add metadata like titles and dates to a single page. This adds a way to
do that: if the ID of a data file (its path without extension) matches
the ID of a page, only propagating data to that page. So
`manual/notebooks.yml` only gets added to `manual/notebooks.ipynb`.